fix posts not loading after redirect
