+++ app/client/src/App.js
@@ -1,5 +1,5 @@
 import { Route, Routes, Navigate } from "react-router-dom";
-import React, { useContext, useEffect } from "react";
+import React, { useContext } from "react";
 import Home from "./pages/Home";
 import PostsSpec from "./pages/PostsSpec";
 import Auth from "./pages/Auth";
@@ -7,33 +7,9 @@ import User from "./pages/User";
 import ProtectedRoute from "./components/ProtectedRoute/ProtectedRoute";
 import Modal from "./components/Modal/Modal";
 import { ModalCtx } from "./features/modal-ctx";
-import { AuthCtx } from "./features/auth-ctx";
-import axios from "axios";
 
 // LOGOUT: Clear cookies
 const App = () => {
-  const authMgr = useContext(AuthCtx);
-
-  const fetchLoggedUser = async () => {
-    await axios
-      .get("/api/logged_user", { withCredentials: true })
-      .then((serverRes) => {
-        authMgr.loginUser(serverRes.data);
-      })
-      .catch((err) => {
-        if (err.response.status == 401) {
-          authMgr.setIsLoggedIn(false);
-          console.log("not logged");
-        } else {
-          console.error(err);
-        }
-      });
-  };
-
-  useEffect(() => {
-    fetchLoggedUser();
-  }, []);
-
   const modalMgr = useContext(ModalCtx);
   return (
     <>
